* Done Notifications
* Working on Split into 2 sections

+++ INB348/GroupsTableViewController.m
@@ -27,7 +27,7 @@
         [query findObjectsInBackgroundWithBlock:^(NSArray *objects, NSError *error) {
             if (!error) {
                 // The find succeeded.
-                NSLog(@"Successfully retrieved %d UserGroups.", objects.count);
+                NSLog(@"Successfully retrieved %lu UserGroups.", (unsigned long)objects.count);
                 
                 // Do something with the found objects
                 self.userGroups = [objects mutableCopy];
